Special AVX2 version for Rng256 and Rng512.
Rng512 is bottlenecked by the memory bandwidth.
Needs Rust 1.57 for const panic.
