+++ src/stable_simd/mod.rs
@@ -1,0 +1,7 @@
+mod rng128;
+mod rng256;
+mod rng512;
+
+pub use rng128::Rng128;
+pub use rng256::Rng256;
+pub use rng512::Rng512;
